pyproject.toml
[tool.poetry]
name = "byte-thread"
version = "0.1.0"
description = ""
authors = ["jztchl <kmvishnu625@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.4"
djangorestframework = "^3.16.0"
pre-commit = "^4.2.0"
django-channels = "^0.7.0"
daphne = "^4.2.1"
channels-redis = "^4.2.1"
djangorestframework-simplejwt = "^5.5.0"
dj-rest-auth = "^7.0.1"
django-allauth = "^65.9.0"
django-extensions = "^4.1"
pillow = "^11.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
